fix: cast double to int

+++ gdms-backend/gdms-server/gdms-server-search/src/main/java/org/graduationdesign/gdmsserversearch/service/MessageService.java
@@ -44,8 +44,8 @@ public class MessageService {
 //            System.out.println(source);
             Message message=new Message();
             message.setTitle((String) source.get("title"));
-            message.setType((int) source.get("type"));
-            message.setId((int) source.get("id"));
+            message.setType(((Double) source.get("type")).intValue());
+            message.setId(((Double) source.get("id")).intValue());
             message.setContent((String) source.get("content"));
             list.add(message);
         }
